Move palette UI drawing code into the UI crate.
